+++ components/index.ts
@@ -1,7 +1,7 @@
 export { cn } from '../lib/utils'
 
-export { Button } from './Button'
-export type { ButtonProps } from './Button'
+export { Button, buttonVariants } from './Button'
+export type { ButtonProps, ButtonVariantProps } from './Button'
 
 export { Alert } from './Alert'
 export type { AlertProps } from './Alert'